+++ src/manifest/index.js
@@ -20,7 +20,7 @@ const manifestInput = {
     homepage_url: 'https://github.com/lowwebtech/low-radical',
     short_name: 'NoAmazon',
 
-    permissions: ['tabs', 'storage', 'webRequest', 'webNavigation', 'webRequestBlocking', '<all_urls>'],
+    permissions: ['tabs', 'storage', 'webRequest',/* 'webNavigation',*/ 'webRequestBlocking', '<all_urls>'],
     content_security_policy: "script-src 'self' 'unsafe-eval'; object-src 'self'",
 
     '__chrome|firefox__author': 'lowwebtech',