+++ src/scripts/comp/degrade.js
@@ -1,6 +1,6 @@
 import browser from 'webextension-polyfill';
 import { formatDotcoms, amazon } from '../data/amazon_urls';
-import { formatedCssRules } from '../data/css_rules';
+import { hiddenCSSRules } from '../data/css_rules';
 
 let isDegrading = false
 
@@ -44,11 +44,11 @@ function onTabUpdate(tabId, changeInfo, tab){
       for (let i = 0, lg = amazon.length; i<lg; i++) {
         if (tab.url.indexOf(amazon[i]) !== -1) {
           console.log('degrade', amazon[i]);
-          let specificRules = getDegradedCSS(amazon[i]);
+          
           let code = `
               * { transition: none !important; animation: none !important; }
               body { filter: grayscale(1); }
-              ${specificRules}
+              ${hiddenCSSRules}
               `;
 
           browser.tabs.insertCSS(tabId, {
@@ -59,15 +59,6 @@ function onTabUpdate(tabId, changeInfo, tab){
     }
   }
 }
-
-function getDegradedCSS( website ){
-  if(formatedCssRules[ website ]){
-    return formatedCssRules[ website ] + '{ display:none!important; }';
-  }else{
-    return '';
-  }
-}
-
 
 function blockAssets(requestDetails) {
   return {